HUM-98 - sign up page refactor
