+++ packages/apps/human-app/frontend/src/modules/worker/utils/formatted-sign-up-error-message.tsx
@@ -1,0 +1,10 @@
+import { t } from 'i18next';
+import { FetchError } from '@/api/fetcher';
+
+export function formattedSignUpErrorMessage(
+  unknownError: unknown
+): string | undefined {
+  if (unknownError instanceof FetchError && unknownError.status === 409) {
+    return t('worker.signUpForm.errors.emailTaken');
+  }
+}
